More float vs int wxPython fixes.

+++ Animation.py
@@ -632,7 +632,7 @@ class Animation(wx.Control):
 		else:
 			tStr = '{}:{:02d}:{:02d}'.format(secs // (60*60), (secs // 60)%60, secs % 60 )
 		tWidth, tHeight = dc.GetTextExtent( tStr )
-		dc.DrawText( tStr, 4*r - tWidth, 2*r - tHeight )
+		dc.DrawText( tStr, int(4*r - tWidth), int(2*r - tHeight) )
 		
 	def OnEraseBackground(self, event):
 		# This is intentionally empty, because we are using the combination
